fix(plugins): Rename dirent.h to tr_dirent.h
